Add base controller to aid in generating IIIF responses
Makes mediatr request and handles responses/headers. Prefactor for
iiif-manifest paths
